Bael 2431 (#6073)
* Adding code to java core and deleting it from java-8

* removing failed unittest

* removing space
