--- core-java-8/src/main/java/com/baeldung/interfaces/Motorcycle.java
@@ -1,10 +1,0 @@
-package com.baeldung.interfaces;
-
-import com.baeldung.interfaces.multiinheritance.Transform;
-
-public class Motorcycle implements Transform {
-    @Override
-    public void transform() {
-        // Implementation
-    }
-}
