--- core-java-8/src/main/java/com/baeldung/interfaces/Truck.java
@@ -1,8 +1,0 @@
-package com.baeldung.interfaces;
-
-public class Truck extends Vehicle {
-    @Override
-    public void transform() {
-        // implementation
-    }
-}
