--- core-java-8/src/main/java/com/baeldung/interfaces/Vehicle.java
@@ -1,6 +1,0 @@
-package com.baeldung.interfaces;
-
-import com.baeldung.interfaces.multiinheritance.Transform;
-
-public abstract class Vehicle implements Transform {
-}
