--- core-java-8/src/main/java/com/baeldung/interfaces/multiinheritance/Vehicle.java
@@ -1,13 +1,0 @@
-package com.baeldung.interfaces.multiinheritance;
-
-public class Vehicle implements Fly, Transform {
-    @Override
-    public void fly() {
-        System.out.println("I can Fly!!");
-    }
-
-    @Override
-    public void transform() {
-        System.out.println("I can Transform!!");
-    }
-}
