--- core-java-8/src/main/java/com/baeldung/interfaces/polymorphysim/Shape.java
@@ -1,9 +1,0 @@
-package com.baeldung.interfaces.polymorphysim;
-
-import com.baeldung.interfaces.HasColor;
-
-public interface Shape extends HasColor {
-
-    public abstract String name();
-    public abstract double area();
-}
