+++ core-java/src/main/java/com/baeldung/interfaces/Box.java
@@ -1,0 +1,5 @@
+package com.baeldung.interfaces;
+
+public interface Box extends HasColor {
+    int getHeight();
+}
