+++ core-java/src/main/java/com/baeldung/interfaces/Employee.java
@@ -11,5 +11,4 @@ public class Employee {
     public void setSalary(double salary) {
         this.salary = salary;
     }
-
 }
