+++ core-java/src/main/java/com/baeldung/interfaces/EmployeeSalaryComparator.java
@@ -6,12 +6,11 @@ public class EmployeeSalaryComparator implements Comparator<Employee> {
 
     @Override
     public int compare(Employee employeeA, Employee employeeB) {
-
-        if(employeeA.getSalary() < employeeB.getSalary()){
+        if (employeeA.getSalary() < employeeB.getSalary()) {
             return -1;
-        }else if(employeeA.getSalary() > employeeB.getSalary()){
+        } else if (employeeA.getSalary() > employeeB.getSalary()) {
             return 1;
-        }else{
+        } else {
             return 0;
         }
     }
