+++ core-java/src/main/java/com/baeldung/interfaces/HasColor.java
@@ -1,5 +1,4 @@
 package com.baeldung.interfaces;
 
 public interface HasColor {
-    public String getColor();
 }
