+++ core-java/src/main/java/com/baeldung/interfaces/multiinheritance/Fly.java
@@ -1,5 +1,6 @@
 package com.baeldung.interfaces.multiinheritance;
 
-public abstract interface Fly{
+public interface Fly {
+
     void fly();
 }
