+++ core-java/src/main/java/com/baeldung/interfaces/multiinheritance/Transform.java
