+++ core-java/src/main/java/com/baeldung/interfaces/multiinheritance/Vehicle.java
@@ -1,0 +1,4 @@
+package com.baeldung.interfaces.multiinheritance;
+
+public abstract class Vehicle implements Transform {
+}
