+++ core-java/src/main/java/com/baeldung/interfaces/polymorphysim/Circle.java
@@ -1,0 +1,9 @@
+package com.baeldung.interfaces.polymorphysim;
+
+public class Circle implements Shape {
+
+    @Override
+    public String name() {
+        return "Circle";
+    }
+}
