+++ core-java/src/main/java/com/baeldung/interfaces/polymorphysim/Shape.java
@@ -1,0 +1,6 @@
+package com.baeldung.interfaces.polymorphysim;
+
+public interface Shape {
+
+    String name();
+}
