+++ core-java/src/main/java/com/baeldung/interfaces/polymorphysim/Square.java
@@ -1,0 +1,9 @@
+package com.baeldung.interfaces.polymorphysim;
+
+public class Square implements Shape {
+
+    @Override
+    public String name() {
+        return "Square";
+    }
+}
